% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{best_beta}
\alias{best_beta}
\title{Best Beta}
\usage{
best_beta(data)
}
\arguments{
\item{data}{vector of data points}
}
\value{
best penalty value
}
\description{
Best parameter for penalty
}
\examples{
best_beta(c(0,1,1,2,2,3))
}
