% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R
\name{cost_gauss}
\alias{cost_gauss}
\title{Cost Gauss}
\usage{
cost_gauss(trajectoire)
}
\arguments{
\item{trajectoire}{vector of trajectoire from data points}
}
\value{
a number
}
\description{
Cost function for Gaussian models
}
\examples{
cost_gauss(c(0,0,1,1,2,2,3,4,5))
}
