% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R
\name{cost_negbin}
\alias{cost_negbin}
\title{Cost Negbin}
\usage{
cost_negbin(v)
}
\arguments{
\item{v}{vector of data points}
}
\value{
the optimal cost value
}
\description{
Cost function for Negbin models
}
\examples{
cost_negbin(c(0,1,2))
cost_negbin(c(1,1,1))
cost_negbin(c(1,1,1,2,2,2))
}
