% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R
\name{cost_poiss}
\alias{cost_poiss}
\title{Cost Poisson}
\usage{
cost_poiss(v)
}
\arguments{
\item{v}{vector of data points}
}
\value{
the optimal cost value
}
\description{
Cost function for Poisson models
}
\examples{
cost_poiss(c(0,0,0))
cost_poiss(c(1,1,1))
cost_poiss(c(1,1,1,2,2,2))
}
