% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPOP.R
\name{downupFPOP}
\alias{downupFPOP}
\title{Functional Pruned Optimal Partitioning 1D}
\usage{
downupFPOP(data, cost = "gauss", beta = best_beta(data), affiche = FALSE)
}
\arguments{
\item{data}{vector of data points}

\item{cost}{a number}

\item{beta}{a number}
}
\value{
a vector of changepoints, global cost
}
\description{
Functional Pruned Optimal Partitioning 1D Algorithm
}
\examples{
downupFPOP(c(rnorm(50, mean = 0, sd = 1), rnorm(50, mean = 100, sd = 1)))
}
