\name{ftse100}
\docType{data}
\alias{ftse100}
\title{
FTSE 100 Daily Returns: 2nd April 1984 -- 13th September 2012
}
\description{
 This dataset gives the daily returns (c_{t+1}/c_t -1) of the UK FTSE 100 index from 2nd April 1984 until the 13th September 2012.
}
\usage{
ftse100
}
\format{
	A matrix of dimension 7187 x 2 where the first column is the Date and the second column is the Daily Return.
}
\source{Yahoo! Finance}

\keyword{datasets}
