% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mle.R
\name{GetMle}
\alias{GetMle}
\title{Estimate Coefficients}
\usage{
GetMle(x, window_size, L, method = "ols")
}
\arguments{
\item{x}{The original data to find change points.}

\item{window_size}{The number of observations each window contains.}

\item{L}{Lag order of the dataset. L>=1}

\item{method}{The method used to estimate coefficients.
Must be one of c("yule-walker", "ols", "mle", "yw").
Default is ols (ordinal least squares).}
}
\value{
x_transformed: The transformed data, which are the estimated coefficients of original data.
}
\description{
Transform N dependent data into approximated independent data (N/window_size) x (L+1).
Computes the estimated coefficients of each window of original data.
}
\examples{
N = 1000
N1 = floor(0.1*N)
N2 = floor(0.3*N)
a1 = c(0.8, -0.3); c1 = 0
a2 = c(-0.5, 0.1); c2 = 0
a3 = c(0.5, -0.5); c3 = 0
x = rep(0,N)
L=2
x[1:L] = rnorm(L)
for (n in (L+1):N){
  if (n <= N1) {
    x[n] = x[(n-1):(n-L)] \%*\% a1 + c1 + rnorm(1)
  } else if (n <= (N1+N2)) {
    x[n] = x[(n-1):(n-L)] \%*\% a2 + c2 + rnorm(1)
  }
  else {
    x[n] = x[(n-1):(n-L)] \%*\% a3 + c3 + rnorm(1)
  }
}
GetMle(x,window_size=100,L=2,method="mle")
GetMle(x,window_size=100,L=2,method="ols")
GetMle(x,window_size=100,L=2,method="yw")
}
\references{
J. Ding, Y. Xiang, L. Shen, and V. Tarokh, \emph{Multiple Change Point Analysis:
Fast Implementation and Strong Consistency}. IEEE Transactions on Signal
Processing, vol. 65, no. 17, pp. 4495-4510, 2017.
}
