\name{HC1}
\docType{data}
\alias{HC1}
\title{
G+C Content in Human Chromosome 1
}
\description{
 This dataset gives the G+C content in 3kb windows along the Human Chromosome from 10Mb to 33Mb (no missing data).
}
\usage{
HC1
}
\format{
	A vector of length 23553.
}
\source{http://www.ncbi.nlm.nih.gov/mapview/map_search.cgi?taxid=9606&build=previous}

\keyword{datasets}
