\name{Lai2005fig3}
\docType{data}
\alias{Lai2005fig3}
\title{
Normalized glioblastoma profile for chromosome 13
}
\description{
This dataset is taken from Lai W, Johnson MJ, Kucherlapati R, Park PJ, Bioinformatics , 2005.  The paper states that the original source of the data is from Bredel et al. (2005).  The data is Chromosome 13 in GBM31.
}
\usage{
Lai2005fig3
}
\format{
	A matrix of dimensions 797 x 5.  The columns are Spot, CH, POS.start, POS.end, GBM31.
}
\source{http://compbio.med.harvard.edu/Supplements/Bioinformatics05b/Profiles/Chrom_13_GBM31.xls}

\keyword{datasets}
