\name{Lai2005fig4}
\docType{data}
\alias{Lai2005fig4}
\title{
Normalized glioblastoma profile for an excerpt of chromosome 7, the EGFR locus.
}
\description{
This dataset is taken from Lai W, Johnson MJ, Kucherlapati R, Park PJ, Bioinformatics , 2005.  The paper states that the original source of the data is from Bredel et al. (2005).  The data is an excerpt of chromosome 7 in GBM29 from 40 to 65 Mb.
}
\usage{
Lai2005fig4
}
\format{
	A matrix of dimensions 193 x 5.  The columns are Spot, CH, POS.start, POS.end, GBM31.
}
\source{http://compbio.med.harvard.edu/Supplements/Bioinformatics05b/Profiles/Chrom_7_from40_to65Mb_GBM29.xls}

\keyword{datasets}
