% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OP.R
\name{myOP}
\alias{myOP}
\title{Optimal Partitioning}
\usage{
myOP(data, cost = "gauss", beta = best_beta(data))
}
\arguments{
\item{data}{vector of data points}

\item{cost}{a number}

\item{beta}{a number}
}
\value{
a vector of changepoints, a number for the complexity (cost of computations)
}
\description{
Optimal Partitioning Algorithm
}
\examples{
myOP(c(0,0,1,1,0,0,0), beta = 0.1)
myOP(c(rnorm(50, mean = 0, sd = 0), rnorm(50, mean = 10, sd = 0)))
}
