% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PELT.R
\name{myPELT}
\alias{myPELT}
\title{Pruned Exact Linear Time}
\usage{
myPELT(data, cost = "gauss", beta = best_beta(data))
}
\arguments{
\item{data}{vector of data points}

\item{cost}{a number}

\item{beta}{a number}
}
\value{
a vector of changepoints, global cost
}
\description{
Pruned Exact Linear Time Algorithm
}
\examples{
myPELT(c(0,0,1,1,0,0,0), beta = 0.00001)
myPELT(c(rnorm(50, mean = 0, sd = 0), rnorm(50, mean = 10, sd = 0)), beta = 1)
}
