\name{wave.c44137}
\docType{data}
\alias{wave.c44137}
\title{
Wave data from buoy c44137
}
\description{
	This dataset gives the significant wave heights from buoy c44137 obtained from the Fisheries and Oceans Canada, East Scotian Slop.  The data are taken at hourly intervals from January 2005 until September 2012.
}
\usage{
wave.c44137
}
\format{
	A vector of length 63651.
}
\source{http://www.meds-sdmm.dfo-mpo.gc.ca/isdm-gdsi/waves-vagues/search-recherche/list-liste/data-donnees-eng.asp?medsid=C44137}

\keyword{datasets}
