% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R
\name{cost_gauss}
\alias{cost_gauss}
\title{Cost Gauss}
\usage{
cost_gauss(v)
}
\arguments{
\item{v}{vector of data points}
}
\value{
the optimal cost value
}
\description{
Cost function for Gaussian models
}
\examples{
cost_gauss(c(0,0,1,1))
}
