% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R
\name{cost_poiss}
\alias{cost_poiss}
\title{Cost Poisson}
\usage{
cost_poiss(trajectoire)
}
\arguments{
\item{trajectoire}{vector of trajectoire from data points}
}
\value{
a number
}
\description{
Cost function for Poisson models
}
\examples{
cost_poiss(c(1,2,3))
cost_poiss(c(1,1,1,2,2,2))
}
