% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OP.R
\name{myOP}
\alias{myOP}
\title{Optimal Partitioning}
\usage{
myOP(data, cost = "gauss", beta = best_beta(data))
}
\arguments{
\item{data}{vector of data points}

\item{cost}{type of my cost (gauss or poisson)}

\item{beta}{penalty value}
}
\value{
a vector of changepoints, a number for the complexity (cost of computations)
}
\description{
Optimal Partitioning Algorithm
}
\examples{
myOP(c(0,0,1,1,0,0,0), beta = 0.00001)
myOP(c(rnorm(50, mean = 0, sd = 0), rnorm(50, mean = 10, sd = 0)))
myOP(data_generator(25, chpts = c(10,20), means = c(20,0,20), type = "gauss"), beta = 5)
}
